refactor(pxe): decompose PXEOracleInterface
Deprecate `PXEOracleInterface` in favor of more granular and explicit dependencies.

I apologize in advance for the length of this PR, the commits show the progression that led to this final form, and prove that the end result didn't come to me OTOH, but it was the result of dozens of stepwise refactors. At the current state it is possible to spin multiple little PRs off this one if that's deemed saner, it just wasn't clear that would be the case when I originally set out to do this.

## New service layer

To replace `PXEOracleInterface` (aka `ExecutionDataProvider`), and at the same time keep oracle classes relatively lean and concerns separated, I'm introducing a service layer on top of the existing providers layer.

Due to this PR being big enough as is, and there being more pressing issues to address, I'm not applying this idea to completion, just using it where it helps me untangle things. I do think going further in this direction would help make everything be more unit testable.

The core rules of thumb behind this (long term) are:

- Providers shouldn't talk to each other. This is a simplicity forcing function. Providers are database abstractions.
- Services can talk to each other, and to the provider(s) they encapsulate. Each provider should only be accessed by one service, but one service can manage multiple data providers. So: services get other read/write data needs met by talking to other services.
- The rest of the app talks to services instead of directly to providers. Again, I'm not tackling that here because the PR is big enough as is, but it's an architecture we could slowly move towards.

## Capsules as plumbing

Something that became clear along the way of this refactor is that capsule state management in many cases is a transport concern, and as such code becomes much more cohesive if we separate the code that deals with that, from the code that actually operates on the contents. Maybe the nicest example of that can be seen in `utilityBulkRetrieveLogs`.

## Working with Anchor blocks

I think most of the places where we now consume `AnchorBlockDataProvider` to get the current anchor block, should be modified to instead receive the anchor block as param, which will make it easier to see that in reality the anchor block is pinned for a given `ContractFunctionSimulator` activation. That, in turn, should make the work on DB integrity easier, and it will also be helpful to eventually adopt a less conservative approach towards concurrency.

## Temporary complexity increase (waiting for new tagging sync work)

There are some parts of the app that might look a bit more convoluted with this refactor, mostly due to porting sync tagging functionality to make things compile and pass tests. Note that @benesjan is overhauling sync tagging, so I'm deliberately not investing effort on simplifying those parts and will instead wait and see what's necessary once that work is available.

Some specific cases of this situation are found on:
- Test setup of `oracle_version_is_checked.test.ts`: it got much more complex given `ContractFunctionSimulator` now receives all the data providers it needs to operate instead of hiding everything under `PXEOracleInterface`. A lot of this complexity is probably derived from the old tag sync'ing logic and I'm expecting it to just disintegrate.
- `private_execution.test.ts`: I'm hoping a lot of the boilerplate will be unnecessary once the new work on tagging lands.
- `LogService` and its tests

Fixes #17776
