--- yarn-project/pxe/src/contract_function_simulator/pxe_oracle_interface.test.ts
@@ -1,1125 +1,0 @@
-import { BlockNumber } from '@aztec/foundation/branded-types';
-import { timesParallel } from '@aztec/foundation/collection';
-import { randomInt } from '@aztec/foundation/crypto/random';
-import { Fq, Fr } from '@aztec/foundation/curves/bn254';
-import { KeyStore } from '@aztec/key-store';
-import { openTmpStore } from '@aztec/kv-store/lmdb-v2';
-import { EventSelector } from '@aztec/stdlib/abi';
-import { AztecAddress } from '@aztec/stdlib/aztec-address';
-import { L2BlockHash, randomDataInBlock } from '@aztec/stdlib/block';
-import { CompleteAddress } from '@aztec/stdlib/contract';
-import { computeUniqueNoteHash, siloNoteHash, siloNullifier, siloPrivateLog } from '@aztec/stdlib/hash';
-import type { AztecNode } from '@aztec/stdlib/interfaces/client';
-import { computeAddress, deriveKeys } from '@aztec/stdlib/keys';
-import { DirectionalAppTaggingSecret, PrivateLog, PublicLog, TxScopedL2Log } from '@aztec/stdlib/logs';
-import { NoteDao, NoteStatus } from '@aztec/stdlib/note';
-import { MerkleTreeId } from '@aztec/stdlib/trees';
-import {
-  BlockHeader,
-  GlobalVariables,
-  type IndexedTxEffect,
-  TxEffect,
-  TxHash,
-  randomIndexedTxEffect,
-} from '@aztec/stdlib/tx';
-
-import { jest } from '@jest/globals';
-import { type MockProxy, mock } from 'jest-mock-extended';
-
-import { AddressDataProvider } from '../storage/address_data_provider/address_data_provider.js';
-import { AnchorBlockDataProvider } from '../storage/anchor_block_data_provider/anchor_block_data_provider.js';
-import { CapsuleDataProvider } from '../storage/capsule_data_provider/capsule_data_provider.js';
-import { ContractDataProvider } from '../storage/contract_data_provider/contract_data_provider.js';
-import { NoteDataProvider } from '../storage/note_data_provider/note_data_provider.js';
-import { PrivateEventDataProvider } from '../storage/private_event_data_provider/private_event_data_provider.js';
-import { RecipientTaggingDataProvider } from '../storage/tagging_data_provider/recipient_tagging_data_provider.js';
-import { SenderTaggingDataProvider } from '../storage/tagging_data_provider/sender_tagging_data_provider.js';
-import { WINDOW_HALF_SIZE } from '../tagging/constants.js';
-import { SiloedTag } from '../tagging/siloed_tag.js';
-import { Tag } from '../tagging/tag.js';
-import { LogRetrievalRequest } from './noir-structs/log_retrieval_request.js';
-import { PXEOracleInterface } from './pxe_oracle_interface.js';
-
-jest.setTimeout(30_000);
-
-async function computeSiloedTagForIndex(
-  sender: { completeAddress: CompleteAddress; ivsk: Fq },
-  recipient: AztecAddress,
-  contractAddress: AztecAddress,
-  index: number,
-) {
-  const secret = await DirectionalAppTaggingSecret.compute(
-    sender.completeAddress,
-    sender.ivsk,
-    recipient,
-    contractAddress,
-    recipient,
-  );
-  const tag = await Tag.compute({ secret, index });
-  return SiloedTag.compute(tag, contractAddress);
-}
-
-describe('PXEOracleInterface', () => {
-  let aztecNode: MockProxy<AztecNode>;
-
-  let addressDataProvider: AddressDataProvider;
-  let privateEventDataProvider: PrivateEventDataProvider;
-  let contractDataProvider: ContractDataProvider;
-  let noteDataProvider: NoteDataProvider;
-  let senderTaggingDataProvider: SenderTaggingDataProvider;
-  let recipientTaggingDataProvider: RecipientTaggingDataProvider;
-  let anchorBlockDataProvider: AnchorBlockDataProvider;
-  let capsuleDataProvider: CapsuleDataProvider;
-  let keyStore: KeyStore;
-
-  let recipient: CompleteAddress;
-  let contractAddress: AztecAddress;
-
-  let pxeOracleInterface: PXEOracleInterface;
-
-  // The block number of the first log to be emitted.
-  const MIN_BLOCK_NUMBER_OF_A_LOG = BlockNumber(1);
-  // The block number of the last log to be emitted.
-  const MAX_BLOCK_NUMBER_OF_A_LOG = BlockNumber(3);
-
-  beforeEach(async () => {
-    const store = await openTmpStore('test');
-    aztecNode = mock<AztecNode>();
-    contractDataProvider = new ContractDataProvider(store);
-    jest.spyOn(contractDataProvider, 'getDebugContractName').mockImplementation(() => Promise.resolve('TestContract'));
-
-    addressDataProvider = new AddressDataProvider(store);
-    privateEventDataProvider = new PrivateEventDataProvider(store);
-    noteDataProvider = await NoteDataProvider.create(store);
-    anchorBlockDataProvider = new AnchorBlockDataProvider(store);
-    senderTaggingDataProvider = new SenderTaggingDataProvider(store);
-    recipientTaggingDataProvider = new RecipientTaggingDataProvider(store);
-    capsuleDataProvider = new CapsuleDataProvider(store);
-    keyStore = new KeyStore(store);
-    pxeOracleInterface = new PXEOracleInterface(
-      aztecNode,
-      keyStore,
-      contractDataProvider,
-      noteDataProvider,
-      capsuleDataProvider,
-      anchorBlockDataProvider,
-      senderTaggingDataProvider,
-      recipientTaggingDataProvider,
-      addressDataProvider,
-      privateEventDataProvider,
-    ); // Set up contract address
-    contractAddress = await AztecAddress.random();
-    // Set up recipient account
-    recipient = await keyStore.addAccount(new Fr(69), Fr.random());
-    await addressDataProvider.addCompleteAddress(recipient);
-
-    // PXEOracleInterface.syncTaggedLogs(...) function syncs logs up to the block number up to which PXE synced. We set
-    // the synced block number to that of the last emitted log to receive all the logs by default.
-    await setSyncedBlockNumber(MAX_BLOCK_NUMBER_OF_A_LOG);
-  });
-
-  describe('sync tagged logs', () => {
-    const NUM_SENDERS = 10;
-
-    let senders: { completeAddress: CompleteAddress; ivsk: Fq; secretKey: Fr }[];
-
-    async function generateMockLogs(tagIndex: number) {
-      const logs: { [k: string]: TxScopedL2Log[] } = {};
-
-      // Add a random note from every address in the address book for our account with index tagIndex
-      // Compute the tag as sender (knowledge of preaddress and ivsk)
-      for (const sender of senders) {
-        const tag = await computeSiloedTagForIndex(sender, recipient.address, contractAddress, tagIndex);
-        const log = new TxScopedL2Log(
-          TxHash.random(),
-          0,
-          0,
-          MIN_BLOCK_NUMBER_OF_A_LOG,
-          L2BlockHash.random(),
-          PrivateLog.random(tag.value),
-        );
-        logs[tag.toString()] = [log];
-      }
-      // Accumulated logs intended for recipient: NUM_SENDERS
-
-      // Add a random note from the first sender in the address book, repeating the tag
-      // Compute the tag as sender (knowledge of preaddress and ivsk)
-      const firstSender = senders[0];
-      const tag = await computeSiloedTagForIndex(firstSender, recipient.address, contractAddress, tagIndex);
-      const log = new TxScopedL2Log(
-        TxHash.random(),
-        1,
-        0,
-        BlockNumber.ZERO,
-        L2BlockHash.random(),
-        PrivateLog.random(tag.value),
-      );
-      logs[tag.toString()].push(log);
-      // Accumulated logs intended for recipient: NUM_SENDERS + 1
-
-      // Add a random note from half the address book for our account with index tagIndex + 1
-      // Compute the tag as sender (knowledge of preaddress and ivsk)
-      for (let i = NUM_SENDERS / 2; i < NUM_SENDERS; i++) {
-        const sender = senders[i];
-        const tag = await computeSiloedTagForIndex(sender, recipient.address, contractAddress, tagIndex + 1);
-        const blockNumber = BlockNumber(2);
-        const log = new TxScopedL2Log(
-          TxHash.random(),
-          0,
-          0,
-          blockNumber,
-          L2BlockHash.random(),
-          PrivateLog.random(tag.value),
-        );
-        logs[tag.toString()] = [log];
-      }
-      // Accumulated logs intended for recipient: NUM_SENDERS + 1 + NUM_SENDERS / 2
-
-      // Add a random note from every address in the address book for a random recipient with index tagIndex
-      // Compute the tag as sender (knowledge of preaddress and ivsk)
-      for (const sender of senders) {
-        const keys = await deriveKeys(Fr.random());
-        const partialAddress = Fr.random();
-        const randomRecipient = await computeAddress(keys.publicKeys, partialAddress);
-        const tag = await computeSiloedTagForIndex(sender, randomRecipient, contractAddress, tagIndex);
-        const log = new TxScopedL2Log(
-          TxHash.random(),
-          0,
-          0,
-          MAX_BLOCK_NUMBER_OF_A_LOG,
-          L2BlockHash.random(),
-          PrivateLog.random(tag.value),
-        );
-        logs[tag.toString()] = [log];
-      }
-      // Accumulated logs intended for recipient: NUM_SENDERS + 1 + NUM_SENDERS / 2
-
-      // Set up the getPrivateLogsByTags mock
-      aztecNode.getLogsByTags.mockImplementation(tags => {
-        return Promise.resolve(tags.map(tag => logs[tag.toString()] ?? []));
-      });
-    }
-
-    // Set to a random value in this test we don't care about Noir loading the logs from the capsule array.
-    const PENDING_TAGGED_LOG_ARRAY_BASE_SLOT = Fr.random();
-
-    beforeEach(async () => {
-      // Set up the address book
-      senders = await timesParallel(NUM_SENDERS, async index => {
-        const keys = await deriveKeys(new Fr(index));
-        const partialAddress = Fr.random();
-        const address = await computeAddress(keys.publicKeys, partialAddress);
-        const completeAddress = await CompleteAddress.create(address, keys.publicKeys, partialAddress);
-        return { completeAddress, ivsk: keys.masterIncomingViewingSecretKey, secretKey: new Fr(index) };
-      });
-      for (const sender of senders) {
-        await recipientTaggingDataProvider.addSenderAddress(sender.completeAddress.address);
-      }
-      aztecNode.getLogsByTags.mockReset();
-      aztecNode.getTxEffect.mockResolvedValue({
-        ...randomDataInBlock(await TxEffect.random({ numNullifiers: 1 })),
-        txIndexInBlock: 0,
-      });
-    });
-
-    it('should sync tagged logs', async () => {
-      const tagIndex = 0;
-      await generateMockLogs(tagIndex);
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // We expect to have all logs intended for the recipient synced (and hence stored in the capsule for later
-      // processing), one per sender + 1 with a duplicated tag for the first sender + half of the logs for the second
-      // index
-      await expectPendingTaggedLogArrayLengthToBe(contractAddress, NUM_SENDERS + 1 + NUM_SENDERS / 2);
-
-      // Recompute the secrets (as recipient) to ensure indexes are updated
-      const ivsk = await keyStore.getMasterIncomingViewingSecretKey(recipient.address);
-      const secrets = await Promise.all(
-        senders.map(sender =>
-          DirectionalAppTaggingSecret.compute(
-            recipient,
-            ivsk,
-            sender.completeAddress.address,
-            contractAddress,
-            recipient.address,
-          ),
-        ),
-      );
-
-      // First sender should have 2 logs, but keep index 0 since they were built using the same tag
-      // Next 4 senders should also have index 0 = offset + 0
-      // Last 5 senders should have index 1 = offset + 1
-      const indexes = await recipientTaggingDataProvider.getLastUsedIndexes(secrets);
-
-      expect(indexes).toHaveLength(NUM_SENDERS);
-      expect(indexes).toEqual([0, 0, 0, 0, 0, 1, 1, 1, 1, 1]);
-
-      // We should have called the node 2 times:
-      // 2 times: first time during initial request, second time after pushing the edge of the window once
-      expect(aztecNode.getLogsByTags.mock.calls.length).toBe(2);
-    });
-
-    it('should sync tagged logs with a sender index offset', async () => {
-      const tagIndex = 5;
-      await generateMockLogs(tagIndex);
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // We expect to have all logs intended for the recipient, one per sender + 1 with a duplicated tag for the first
-      // one + half of the logs for the second index
-      await expectPendingTaggedLogArrayLengthToBe(contractAddress, NUM_SENDERS + 1 + NUM_SENDERS / 2);
-
-      // Recompute the secrets (as recipient) to ensure indexes are updated
-      const ivsk = await keyStore.getMasterIncomingViewingSecretKey(recipient.address);
-      const secrets = await Promise.all(
-        senders.map(sender =>
-          DirectionalAppTaggingSecret.compute(
-            recipient,
-            ivsk,
-            sender.completeAddress.address,
-            contractAddress,
-            recipient.address,
-          ),
-        ),
-      );
-
-      // First sender should have 2 logs, but keep index 5 since they were built using the same tag
-      // Next 4 senders should also have index 5 = offset
-      // Last 5 senders should have index 6 = offset + 1
-      const indexes = await recipientTaggingDataProvider.getLastUsedIndexes(secrets);
-
-      expect(indexes).toHaveLength(NUM_SENDERS);
-      expect(indexes).toEqual([5, 5, 5, 5, 5, 6, 6, 6, 6, 6]);
-
-      // We should have called the node 2 times:
-      // 2 times: first time during initial request, second time after pushing the edge of the window once
-      expect(aztecNode.getLogsByTags.mock.calls.length).toBe(2);
-    });
-
-    it("should sync tagged logs for which indexes are not updated if they're inside the window", async () => {
-      const tagIndex = 1;
-      await generateMockLogs(tagIndex);
-
-      // Recompute the secrets (as recipient) to update indexes
-      const ivsk = await keyStore.getMasterIncomingViewingSecretKey(recipient.address);
-      const secrets = await Promise.all(
-        senders.map(sender =>
-          DirectionalAppTaggingSecret.compute(
-            recipient,
-            ivsk,
-            sender.completeAddress.address,
-            contractAddress,
-            recipient.address,
-          ),
-        ),
-      );
-
-      // Set last used indexes to 1 (so next scan starts at 2)
-      await recipientTaggingDataProvider.setLastUsedIndexes(secrets.map(secret => ({ secret, index: 1 })));
-
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // Even if our index as recipient is higher than what the sender sent, we should be able to find the logs
-      // since the window starts at Math.max(0, 2 - window_size) = 0
-      await expectPendingTaggedLogArrayLengthToBe(contractAddress, NUM_SENDERS + 1 + NUM_SENDERS / 2);
-
-      // First sender should have 2 logs, but keep index 1 since they were built using the same tag
-      // Next 4 senders should also have index 1 = tagIndex
-      // Last 5 senders should have index 2 = tagIndex + 1
-      const indexes = await recipientTaggingDataProvider.getLastUsedIndexes(secrets);
-
-      expect(indexes).toHaveLength(NUM_SENDERS);
-      expect(indexes).toEqual([1, 1, 1, 1, 1, 2, 2, 2, 2, 2]);
-
-      // We should have called the node 2 times:
-      // first time during initial request, second time after pushing the edge of the window once
-      expect(aztecNode.getLogsByTags.mock.calls.length).toBe(2);
-    });
-
-    it("should not sync tagged logs for which indexes are not updated if they're outside the window", async () => {
-      const tagIndex = 0;
-      await generateMockLogs(tagIndex);
-
-      // Recompute the secrets (as recipient) to update indexes
-      const ivsk = await keyStore.getMasterIncomingViewingSecretKey(recipient.address);
-      const secrets = await Promise.all(
-        senders.map(sender =>
-          DirectionalAppTaggingSecret.compute(
-            recipient,
-            ivsk,
-            sender.completeAddress.address,
-            contractAddress,
-            recipient.address,
-          ),
-        ),
-      );
-
-      // We set the last used indexes to WINDOW_HALF_SIZE so that next scan starts at WINDOW_HALF_SIZE + 1,
-      // which is outside the window, and for this reason no updates should be triggered.
-      const index = WINDOW_HALF_SIZE + 1;
-      await recipientTaggingDataProvider.setLastUsedIndexes(secrets.map(secret => ({ secret, index })));
-
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // Only half of the logs should be synced since we start from index 1 = (11 - window_size), the other half should
-      // be skipped
-      await expectPendingTaggedLogArrayLengthToBe(contractAddress, NUM_SENDERS / 2);
-
-      // Indexes should remain where we set them (window_size)
-      const indexes = await recipientTaggingDataProvider.getLastUsedIndexes(secrets);
-
-      expect(indexes).toHaveLength(NUM_SENDERS);
-      expect(indexes).toEqual([index, index, index, index, index, index, index, index, index, index]);
-
-      // We should have called the node once and that is only for the first window
-      expect(aztecNode.getLogsByTags.mock.calls.length).toBe(1);
-    });
-
-    it('should sync tagged logs from scratch after a DB wipe', async () => {
-      const tagIndex = 0;
-      await generateMockLogs(tagIndex);
-
-      // Recompute the secrets (as recipient) to update indexes
-      const ivsk = await keyStore.getMasterIncomingViewingSecretKey(recipient.address);
-      const secrets = await Promise.all(
-        senders.map(sender =>
-          DirectionalAppTaggingSecret.compute(
-            recipient,
-            ivsk,
-            sender.completeAddress.address,
-            contractAddress,
-            recipient.address,
-          ),
-        ),
-      );
-
-      await recipientTaggingDataProvider.setLastUsedIndexes(
-        secrets.map(secret => ({ secret, index: WINDOW_HALF_SIZE + 2 })),
-      );
-
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // No logs should be synced (and hence no capsules stored) since we start from index 2 = 12 - window_size
-      await expectPendingTaggedLogArrayLengthToBe(contractAddress, 0);
-
-      // Since no logs were synced, window edge hash not been pushed and for this reason we should have called
-      // the node only once for the initial window
-      expect(aztecNode.getLogsByTags.mock.calls.length).toBe(1);
-
-      aztecNode.getLogsByTags.mockClear();
-
-      // Wipe the database
-      await recipientTaggingDataProvider.resetNoteSyncData();
-
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // First sender should have 2 logs, but keep index 0 since they were built using the same tag
-      // Next 4 senders should also have index 0 = offset
-      // Last 5 senders should have index 1 = offset + 1
-      const indexes = await recipientTaggingDataProvider.getLastUsedIndexes(secrets);
-
-      expect(indexes).toHaveLength(NUM_SENDERS);
-      expect(indexes).toEqual([0, 0, 0, 0, 0, 1, 1, 1, 1, 1]);
-
-      // We should have called the node 2 times:
-      // first time during initial request, second time after pushing the edge of the window once
-      expect(aztecNode.getLogsByTags.mock.calls.length).toBe(2);
-    });
-
-    it('should not sync tagged logs with a blockNumber larger than the block number to which PXE is synced', async () => {
-      // We set the block number to which PXE is synced to a block number in which only the first batch of logs was
-      // emitted and then we check that we receive logs only from this batch.
-      await setSyncedBlockNumber(MIN_BLOCK_NUMBER_OF_A_LOG);
-
-      const tagIndex = 0;
-      await generateMockLogs(tagIndex);
-      await pxeOracleInterface.syncTaggedLogs(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-
-      // Only NUM_SENDERS + 1 logs should be synched, since the rest have blockNumber > 1
-      await expectPendingTaggedLogArrayLengthToBe(contractAddress, NUM_SENDERS + 1);
-    });
-
-    const expectPendingTaggedLogArrayLengthToBe = async (contractAddress: AztecAddress, expectedLength: number) => {
-      // Capsule array length is stored in the array base slot.
-      const capsule = await capsuleDataProvider.loadCapsule(contractAddress, PENDING_TAGGED_LOG_ARRAY_BASE_SLOT);
-      if (expectedLength === 0 && capsule === null) {
-        // If expected length is 0 we are fine with the capsule not existing since the array might not have been
-        // initialized yet.
-        return;
-      }
-      expect(capsule).toBeDefined();
-      expect(capsule!.length).toBe(1);
-      expect(capsule![0].toNumber()).toBe(expectedLength);
-    };
-  });
-
-  describe('deliverEvent', () => {
-    let blockNumber: BlockNumber;
-    let eventSelector: EventSelector;
-    let eventContent: Fr[];
-    let eventCommitment: Fr;
-    let eventNullifier: Fr;
-    let txEffect: TxEffect;
-    let indexedTxEffect: IndexedTxEffect;
-
-    // beforeEach sets up the happy path case, so error modes are tested
-    // by minimally failing happy path conditions
-    beforeEach(async () => {
-      blockNumber = BlockNumber(42);
-      eventSelector = EventSelector.random();
-      eventContent = [Fr.random(), Fr.random()];
-
-      eventCommitment = Fr.random();
-      eventNullifier = await siloNullifier(contractAddress, eventCommitment);
-
-      txEffect = TxEffect.from({
-        ...(await TxEffect.random()),
-        nullifiers: [eventNullifier],
-      });
-
-      indexedTxEffect = {
-        l2BlockNumber: blockNumber,
-        l2BlockHash: L2BlockHash.random(),
-        data: txEffect,
-        txIndexInBlock: 0,
-      };
-
-      /* Happy path context conditions:
-       ** - PXE is sync'd to _at least_ block including tx
-       ** - Node knows tx effect
-       ** - Node knows siloed event commitment
-       */
-      await setSyncedBlockNumber(blockNumber);
-
-      aztecNode.getTxEffect.mockImplementation(() => Promise.resolve(indexedTxEffect));
-
-      aztecNode.findLeavesIndexes.mockImplementation(() =>
-        Promise.resolve([
-          {
-            data: BigInt(0),
-            l2BlockNumber: indexedTxEffect.l2BlockNumber,
-            l2BlockHash: indexedTxEffect.l2BlockHash,
-          },
-        ]),
-      );
-    });
-
-    function deliverEvent(
-      overrides: {
-        eventCommitment?: Fr;
-      } = {},
-    ) {
-      return pxeOracleInterface.deliverEvent(
-        contractAddress,
-        eventSelector,
-        eventContent,
-        overrides.eventCommitment || eventCommitment,
-        txEffect.txHash,
-        recipient.address,
-      );
-    }
-
-    it('should throw when tx does not exist or has no effects', async () => {
-      aztecNode.getTxEffect.mockImplementation(() => Promise.resolve(undefined));
-      await expect(deliverEvent).rejects.toThrow(/Could not find tx effect for tx hash/);
-    });
-
-    it('should throw when tx block has not yet been synchronized', async () => {
-      indexedTxEffect = {
-        ...indexedTxEffect,
-        l2BlockNumber: BlockNumber(blockNumber + 1),
-      };
-      aztecNode.getTxEffect.mockImplementation(() => Promise.resolve(indexedTxEffect));
-
-      await expect(deliverEvent).rejects.toThrow(/Could not find tx effect for tx hash .* as of block number/);
-    });
-
-    it('should throw if event is not in tx effects', async () => {
-      await expect(deliverEvent({ eventCommitment: Fr.random() })).rejects.toThrow(
-        /Event commitment .* is not present in tx/,
-      );
-    });
-
-    it('should throw if event is not in nullifiers', async () => {
-      aztecNode.findLeavesIndexes.mockImplementation(() => Promise.resolve([]));
-
-      await expect(deliverEvent).rejects.toThrow(/Event commitment .* is not present on the nullifier tree/);
-    });
-
-    it('should store event for later retrieval', async () => {
-      await deliverEvent();
-
-      // I should be able to retrieve the private event I just saved using getPrivateEvents
-      const result = await privateEventDataProvider.getPrivateEvents(eventSelector, {
-        contractAddress,
-        fromBlock: blockNumber,
-        toBlock: blockNumber + 1,
-        scopes: [recipient.address],
-      });
-
-      expect(result.length).toEqual(1);
-      expect(result[0].packedEvent).toEqual(eventContent);
-    });
-  });
-
-  describe('deliverNote', () => {
-    // Recipient is different from the owner because recipient refers to the
-    // recipient of the message containing the note, while owner refers to the
-    // owner of the note.
-    let owner: AztecAddress;
-    let storageSlot: Fr;
-    let randomness: Fr;
-    let noteNonce: Fr;
-    let content: Fr[];
-
-    let noteHash: Fr;
-    let uniqueNoteHash: Fr;
-    let nullifier: Fr;
-    let siloedNullifier: Fr;
-
-    let txHash: TxHash;
-    let txEffect: TxEffect;
-    let indexedTxEffect: IndexedTxEffect;
-    let blockNumber: BlockNumber;
-
-    let nullified = false;
-
-    // beforeEach sets up the happy path case, so error modes are tested
-    // by minimally failing happy path conditions
-    beforeEach(async () => {
-      noteHash = Fr.random();
-      nullifier = Fr.random();
-      txHash = TxHash.random();
-      owner = await AztecAddress.random();
-      storageSlot = Fr.random();
-      randomness = Fr.random();
-      noteNonce = Fr.random();
-      content = [Fr.random(), Fr.random()];
-
-      uniqueNoteHash = await computeUniqueNoteHash(noteNonce, await siloNoteHash(contractAddress, noteHash));
-      siloedNullifier = await siloNullifier(contractAddress, nullifier);
-
-      blockNumber = BlockNumber(42);
-
-      txEffect = TxEffect.from({
-        ...(await TxEffect.random()),
-        noteHashes: [uniqueNoteHash],
-      });
-
-      indexedTxEffect = {
-        l2BlockNumber: blockNumber,
-        l2BlockHash: L2BlockHash.random(),
-        data: txEffect,
-        txIndexInBlock: 0,
-      };
-
-      /* Happy path context conditions:
-       ** - PXE is sync'd to _at least_ block including tx
-       ** - Node knows tx effect
-       ** - Node knows unique note hash (and siloed nullifier if requested)
-       */
-      await setSyncedBlockNumber(blockNumber);
-
-      aztecNode.getTxEffect.mockImplementation(queryTxHash =>
-        Promise.resolve(queryTxHash == txHash ? indexedTxEffect : undefined),
-      );
-
-      aztecNode.findLeavesIndexes.mockImplementation((queryBlockNum, treeId, leaves) => {
-        if (queryBlockNum != blockNumber) {
-          throw new Error(`Got a tree query for block ${queryBlockNum} but synced block is ${blockNumber}`);
-        }
-
-        if (treeId == MerkleTreeId.NOTE_HASH_TREE && leaves[0].equals(uniqueNoteHash)) {
-          return Promise.resolve([
-            {
-              data: BigInt(0),
-              l2BlockNumber: indexedTxEffect.l2BlockNumber,
-              l2BlockHash: indexedTxEffect.l2BlockHash,
-            },
-          ]);
-        } else if (treeId == MerkleTreeId.NULLIFIER_TREE && leaves[0].equals(siloedNullifier)) {
-          // Note that returning undefined (i.e. the un-nullified case) covers both scenarios where the note has not
-          // been nullified and where the nullifier is in a block past the synced block.
-          return Promise.resolve([
-            nullified
-              ? {
-                  data: BigInt(0),
-                  l2BlockNumber: indexedTxEffect.l2BlockNumber,
-                  l2BlockHash: indexedTxEffect.l2BlockHash,
-                }
-              : undefined,
-          ]);
-        } else {
-          throw new Error();
-        }
-      });
-    });
-
-    it('should store note if it exists in note hash tree and is not nullified', async () => {
-      await pxeOracleInterface.deliverNote(
-        contractAddress,
-        owner,
-        storageSlot,
-        randomness,
-        noteNonce,
-        content,
-        noteHash,
-        nullifier,
-        txHash,
-        recipient.address,
-      );
-
-      // Verify note was stored
-      const notes = await noteDataProvider.getNotes({ contractAddress, scopes: [recipient.address] });
-
-      expect(notes).toHaveLength(1);
-      expect(notes[0].noteHash.equals(noteHash)).toBe(true);
-    });
-
-    it('should throw if tx hash does not exist', async () => {
-      await expect(
-        pxeOracleInterface.deliverNote(
-          contractAddress,
-          owner,
-          storageSlot,
-          randomness,
-          noteNonce,
-          content,
-          noteHash,
-          nullifier,
-          TxHash.random(),
-          recipient.address,
-        ),
-      ).rejects.toThrow(/Could not find tx effect/);
-    });
-
-    it('should throw if note was not emitted in the tx', async () => {
-      await expect(
-        pxeOracleInterface.deliverNote(
-          contractAddress,
-          owner,
-          storageSlot,
-          randomness,
-          noteNonce,
-          content,
-          Fr.random(), // note hash
-          nullifier,
-          txHash,
-          recipient.address,
-        ),
-      ).rejects.toThrow(/is not present in tx/);
-    });
-
-    it('should throw if tx was mined after synced block number', async () => {
-      await setSyncedBlockNumber(BlockNumber(blockNumber - 1));
-
-      await expect(
-        pxeOracleInterface.deliverNote(
-          contractAddress,
-          owner,
-          storageSlot,
-          randomness,
-          noteNonce,
-          content,
-          noteHash,
-          nullifier,
-          txHash,
-          recipient.address,
-        ),
-      ).rejects.toThrow(/as of block number/);
-    });
-
-    it('should store and immediately remove note if it is already nullified', async () => {
-      nullified = true;
-
-      await pxeOracleInterface.deliverNote(
-        contractAddress,
-        owner,
-        storageSlot,
-        randomness,
-        noteNonce,
-        content,
-        noteHash,
-        nullifier,
-        txHash,
-        recipient.address,
-      );
-
-      // Verify note was removed
-      const notes = await noteDataProvider.getNotes({ contractAddress, scopes: [recipient.address] });
-      expect(notes).toHaveLength(0);
-    });
-  });
-
-  describe('utilityBulkRetrieveLogs', () => {
-    const unsiloedTag = Fr.random();
-    const REQUEST_SLOT = Fr.random();
-    const RESPONSE_SLOT = Fr.random();
-
-    beforeEach(() => {
-      aztecNode.getLogsByTags.mockReset();
-      aztecNode.getTxEffect.mockReset();
-    });
-
-    it('returns no logs if none are found', async () => {
-      aztecNode.getLogsByTags.mockResolvedValue([[]]);
-
-      const request = new LogRetrievalRequest(contractAddress, unsiloedTag);
-
-      await capsuleDataProvider.setCapsuleArray(contractAddress, REQUEST_SLOT, [request.toFields()]);
-      await pxeOracleInterface.bulkRetrieveLogs(contractAddress, REQUEST_SLOT, RESPONSE_SLOT);
-
-      expect((await capsuleDataProvider.readCapsuleArray(contractAddress, REQUEST_SLOT)).length).toEqual(0);
-
-      const responses = await capsuleDataProvider.readCapsuleArray(contractAddress, RESPONSE_SLOT);
-      expect(responses.length).toEqual(1);
-
-      // Check Option::none
-      expect(responses[0][0]).toEqual(new Fr(0)); // TODO: deserialize into option and check properly
-    });
-
-    it('returns a public log if one is found', async () => {
-      const scopedLog = await TxScopedL2Log.random(true);
-      (scopedLog.log as PublicLog).contractAddress = contractAddress;
-
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog]]);
-      const indexedTxEffect = await randomIndexedTxEffect();
-
-      aztecNode.getTxEffect.mockImplementation((txHash: TxHash) =>
-        txHash.equals(scopedLog.txHash) ? Promise.resolve(indexedTxEffect) : Promise.resolve(undefined),
-      );
-
-      const request = new LogRetrievalRequest(contractAddress, scopedLog.log.fields[0]);
-
-      await capsuleDataProvider.setCapsuleArray(contractAddress, REQUEST_SLOT, [request.toFields()]);
-      await pxeOracleInterface.bulkRetrieveLogs(contractAddress, REQUEST_SLOT, RESPONSE_SLOT);
-
-      expect((await capsuleDataProvider.readCapsuleArray(contractAddress, REQUEST_SLOT)).length).toEqual(0);
-
-      const responses = await capsuleDataProvider.readCapsuleArray(contractAddress, RESPONSE_SLOT);
-      expect(responses.length).toEqual(1);
-
-      // Check Option::some
-      expect(responses[0][0]).toEqual(new Fr(1)); // TODO: deserialize into option and check properly
-    });
-
-    it('returns a private log if one is found', async () => {
-      const scopedLog = await TxScopedL2Log.random(false);
-      scopedLog.log.fields[0] = await siloPrivateLog(contractAddress, Fr.random());
-
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog]]);
-      const indexedTxEffect = await randomIndexedTxEffect();
-      aztecNode.getTxEffect.mockResolvedValue(indexedTxEffect);
-
-      aztecNode.getTxEffect.mockImplementation((txHash: TxHash) =>
-        txHash.equals(scopedLog.txHash) ? Promise.resolve(indexedTxEffect) : Promise.resolve(undefined),
-      );
-
-      const request = new LogRetrievalRequest(contractAddress, scopedLog.log.fields[0]);
-
-      await capsuleDataProvider.setCapsuleArray(contractAddress, REQUEST_SLOT, [request.toFields()]);
-      await pxeOracleInterface.bulkRetrieveLogs(contractAddress, REQUEST_SLOT, RESPONSE_SLOT);
-
-      expect((await capsuleDataProvider.readCapsuleArray(contractAddress, REQUEST_SLOT)).length).toEqual(0);
-
-      const responses = await capsuleDataProvider.readCapsuleArray(contractAddress, RESPONSE_SLOT);
-      expect(responses.length).toEqual(1);
-
-      // Check Option::some
-      expect(responses[0][0]).toEqual(new Fr(1)); // TODO: deserialize into option and check properly
-    });
-  });
-
-  describe('getPublicLogByTag', () => {
-    const tag = Fr.random();
-
-    beforeEach(() => {
-      aztecNode.getLogsByTags.mockReset();
-      aztecNode.getTxEffect.mockReset();
-    });
-
-    it('returns null if no logs found for tag', async () => {
-      aztecNode.getLogsByTags.mockResolvedValue([[]]);
-
-      const result = await pxeOracleInterface.getPublicLogByTag(tag, contractAddress);
-      expect(result).toBeNull();
-    });
-
-    it('returns log data when single log found', async () => {
-      const scopedLog = await TxScopedL2Log.random(true);
-      const logContractAddress = (scopedLog.log as PublicLog).contractAddress;
-
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog]]);
-      const indexedTxEffect = await randomIndexedTxEffect();
-      aztecNode.getTxEffect.mockImplementation((txHash: TxHash) =>
-        txHash.equals(scopedLog.txHash) ? Promise.resolve(indexedTxEffect) : Promise.resolve(undefined),
-      );
-
-      const result = (await pxeOracleInterface.getPublicLogByTag(tag, logContractAddress))!;
-
-      expect(result.logPayload).toEqual(scopedLog.log.getEmittedFieldsWithoutTag());
-      expect(result.uniqueNoteHashesInTx).toEqual(indexedTxEffect.data.noteHashes);
-      expect(result.txHash).toEqual(scopedLog.txHash);
-      expect(result.firstNullifierInTx).toEqual(indexedTxEffect.data.nullifiers[0]);
-
-      expect(aztecNode.getLogsByTags).toHaveBeenCalledWith([tag]);
-      expect(aztecNode.getTxEffect).toHaveBeenCalledWith(scopedLog.txHash);
-    });
-
-    it('throws if multiple logs found for tag', async () => {
-      const scopedLog = await TxScopedL2Log.random(true);
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog, scopedLog]]);
-      const logContractAddress = (scopedLog.log as PublicLog).contractAddress;
-
-      await expect(pxeOracleInterface.getPublicLogByTag(tag, logContractAddress)).rejects.toThrow(/Got 2 logs for tag/);
-    });
-
-    it('throws if tx effect not found', async () => {
-      const scopedLog = await TxScopedL2Log.random(true);
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog]]);
-      aztecNode.getTxEffect.mockResolvedValue(undefined);
-      const logContractAddress = (scopedLog.log as PublicLog).contractAddress;
-
-      await expect(pxeOracleInterface.getPublicLogByTag(tag, logContractAddress)).rejects.toThrow(
-        /failed to retrieve tx effects/,
-      );
-    });
-
-    it('returns log fields that are actually emitted', async () => {
-      const logContractAddress = await AztecAddress.random();
-      const logPlaintext = [Fr.random()];
-      const logContent = [tag, ...logPlaintext];
-
-      const log = PublicLog.from({
-        contractAddress: logContractAddress,
-        fields: logContent,
-      });
-      const scopedLogWithPadding = new TxScopedL2Log(
-        TxHash.random(),
-        randomInt(100),
-        randomInt(100),
-        BlockNumber(randomInt(100)),
-        L2BlockHash.random(),
-        log,
-      );
-
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLogWithPadding]]);
-      aztecNode.getTxEffect.mockResolvedValue(await randomIndexedTxEffect());
-
-      const result = await pxeOracleInterface.getPublicLogByTag(tag, logContractAddress);
-
-      expect(result?.logPayload).toEqual(logPlaintext);
-    });
-  });
-
-  describe('getPrivateLogByTag', () => {
-    let tag: Fr;
-
-    beforeEach(() => {
-      tag = Fr.random();
-    });
-
-    it('returns null if no logs found', async () => {
-      aztecNode.getLogsByTags.mockResolvedValue([[]]);
-      const result = await pxeOracleInterface.getPrivateLogByTag(tag);
-      expect(result).toBeNull();
-    });
-
-    it('returns log and tx effect if single log found', async () => {
-      const scopedLog = await TxScopedL2Log.random(false);
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog]]);
-      const indexedTxEffect = await randomIndexedTxEffect();
-      aztecNode.getTxEffect.mockResolvedValue(indexedTxEffect);
-
-      const result = await pxeOracleInterface.getPrivateLogByTag(tag);
-
-      expect(result?.logPayload).toEqual(scopedLog.log.getEmittedFieldsWithoutTag());
-      expect(result?.uniqueNoteHashesInTx).toEqual(indexedTxEffect.data.noteHashes);
-      expect(result?.txHash).toEqual(scopedLog.txHash);
-      expect(result?.firstNullifierInTx).toEqual(indexedTxEffect.data.nullifiers[0]);
-      expect(aztecNode.getTxEffect).toHaveBeenCalledWith(scopedLog.txHash);
-    });
-
-    it('throws if multiple logs found for tag', async () => {
-      const scopedLog = await TxScopedL2Log.random(false);
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog, scopedLog]]);
-
-      await expect(pxeOracleInterface.getPrivateLogByTag(tag)).rejects.toThrow(/Got 2 logs for tag/);
-    });
-
-    it('throws if tx effect not found', async () => {
-      const scopedLog = await TxScopedL2Log.random(false);
-      aztecNode.getLogsByTags.mockResolvedValue([[scopedLog]]);
-      aztecNode.getTxEffect.mockResolvedValue(undefined);
-
-      await expect(pxeOracleInterface.getPrivateLogByTag(tag)).rejects.toThrow(/failed to retrieve tx effects/);
-    });
-  });
-
-  describe('syncNoteNullifiers', () => {
-    let recipient: AztecAddress;
-
-    beforeEach(async () => {
-      // Check that there are no notes in the database
-      const notes = await noteDataProvider.getNotes({ contractAddress });
-      expect(notes).toHaveLength(0);
-
-      // Check that the expected number of accounts is present
-      const accounts = await keyStore.getAccounts();
-      expect(accounts).toHaveLength(1);
-
-      recipient = accounts[0];
-    });
-
-    it('should remove notes that have been nullified', async () => {
-      // Set up initial state with a note
-      const noteDao = await NoteDao.random({ contractAddress });
-
-      // Spy on the noteDataProvider.applyNullifiers to later on have additional guarantee that we really removed
-      // the note.
-      jest.spyOn(noteDataProvider, 'applyNullifiers');
-
-      // Add the note to storage
-      await noteDataProvider.addNotes([noteDao], recipient);
-
-      // Set up the nullifier in the merkle tree
-      const nullifierIndex = randomDataInBlock(123n);
-      aztecNode.findLeavesIndexes.mockResolvedValue([nullifierIndex]);
-
-      // Call the function under test
-      await pxeOracleInterface.syncNoteNullifiers(contractAddress);
-
-      // Verify the note was removed by checking storage
-      const remainingNotes = await noteDataProvider.getNotes({
-        contractAddress,
-        status: NoteStatus.ACTIVE,
-        scopes: [recipient],
-      });
-      expect(remainingNotes).toHaveLength(0);
-
-      // Verify the note was removed by checking the spy
-      expect(noteDataProvider.applyNullifiers).toHaveBeenCalledTimes(1);
-    });
-
-    it('should keep notes that have not been nullified', async () => {
-      // Set up initial state with a note
-      const noteDao = await NoteDao.random({ contractAddress });
-
-      // Add the note to storage
-      await noteDataProvider.addNotes([noteDao], recipient);
-
-      // No nullifier found in merkle tree
-      aztecNode.findLeavesIndexes.mockResolvedValue([undefined]);
-
-      // Call the function under test
-      await pxeOracleInterface.syncNoteNullifiers(contractAddress);
-
-      // Verify note still exists
-      const remainingNotes = await noteDataProvider.getNotes({
-        contractAddress,
-        status: NoteStatus.ACTIVE,
-        scopes: [recipient],
-      });
-      expect(remainingNotes).toHaveLength(1);
-      expect(remainingNotes[0]).toEqual(noteDao);
-    });
-
-    // Verifies that notes are not marked as nullified when their nullifier only exists in blocks that haven't been
-    // synced yet. We mock the nullifier to only exist in blocks beyond our current sync point, then verify the note
-    // is not removed by applyNullifiers.
-    it('should not remove notes if nullifier is in unsynced blocks', async () => {
-      // Set up initial state with a note
-      const noteDao = await NoteDao.random({ contractAddress });
-      const syncedBlockNumber = 100;
-      await setSyncedBlockNumber(BlockNumber(syncedBlockNumber));
-
-      // Add the note to storage
-      await noteDataProvider.addNotes([noteDao], recipient);
-
-      // Mock nullifier to only exist after synced block
-      aztecNode.findLeavesIndexes.mockImplementation(blockNum => {
-        if (typeof blockNum === 'number' && blockNum > syncedBlockNumber) {
-          return Promise.resolve([randomDataInBlock(0n)]);
-        }
-        return Promise.resolve([undefined]);
-      });
-
-      // Call the function under test
-      await pxeOracleInterface.syncNoteNullifiers(contractAddress);
-
-      // Verify note still exists
-      const remainingNotes = await noteDataProvider.getNotes({
-        contractAddress,
-        status: NoteStatus.ACTIVE,
-        scopes: [recipient],
-      });
-      expect(remainingNotes).toHaveLength(1);
-      expect(remainingNotes[0]).toEqual(noteDao);
-    });
-
-    it('should search for notes from all accounts', async () => {
-      // Add multiple accounts to keystore
-      await keyStore.addAccount(Fr.random(), Fr.random());
-      await keyStore.addAccount(Fr.random(), Fr.random());
-
-      expect(await keyStore.getAccounts()).toHaveLength(3);
-
-      // Spy on the noteDataProvider.getNotesSpy
-      const getNotesSpy = jest.spyOn(noteDataProvider, 'getNotes');
-
-      // Call the function under test
-      await pxeOracleInterface.syncNoteNullifiers(contractAddress);
-
-      // Verify applyNullifiers was called once for all accounts
-      expect(getNotesSpy).toHaveBeenCalledTimes(1);
-
-      // Verify getNotes was called with the correct contract address
-      expect(getNotesSpy).toHaveBeenCalledWith(expect.objectContaining({ contractAddress }));
-    });
-  });
-
-  describe('Respects synced block number', () => {
-    const syncedBlockNumber = 100;
-    let contractAddress: AztecAddress;
-    let nullifier: Fr;
-    let leafSlot: Fr;
-
-    beforeEach(async () => {
-      contractAddress = await AztecAddress.random();
-      nullifier = Fr.random();
-      leafSlot = Fr.random();
-      await setSyncedBlockNumber(BlockNumber(syncedBlockNumber));
-    });
-
-    it('throws when getting low nullifier membership witness for future block', async () => {
-      await expect(
-        pxeOracleInterface.getLowNullifierMembershipWitness(BlockNumber(syncedBlockNumber + 1), nullifier),
-      ).rejects.toThrow(`Block number ${syncedBlockNumber + 1} is higher than current block ${syncedBlockNumber}`);
-    });
-
-    it('throws when getting block for future block number', async () => {
-      await expect(pxeOracleInterface.getBlock(BlockNumber(syncedBlockNumber + 1))).rejects.toThrow(
-        `Block number ${syncedBlockNumber + 1} is higher than current block ${syncedBlockNumber}`,
-      );
-    });
-
-    it('throws when getting public data witness for future block', async () => {
-      await expect(
-        pxeOracleInterface.getPublicDataWitness(BlockNumber(syncedBlockNumber + 1), leafSlot),
-      ).rejects.toThrow(`Block number ${syncedBlockNumber + 1} is higher than current block ${syncedBlockNumber}`);
-    });
-
-    it('throws when getting public storage for future block', async () => {
-      await expect(
-        pxeOracleInterface.getPublicStorageAt(BlockNumber(syncedBlockNumber + 1), contractAddress, leafSlot),
-      ).rejects.toThrow(`Block number ${syncedBlockNumber + 1} is higher than current block ${syncedBlockNumber}`);
-    });
-  });
-
-  const setSyncedBlockNumber = (blockNumber: BlockNumber) => {
-    return anchorBlockDataProvider.setHeader(
-      BlockHeader.empty({
-        globalVariables: GlobalVariables.empty({ blockNumber }),
-      }),
-    );
-  };
-});
