+++ yarn-project/pxe/src/entrypoints/client/bundle/index.ts
@@ -3,5 +3,4 @@ export * from '../../../config/index.js';
 export * from '../../../error_enriching.js';
 export * from '../../../storage/index.js';
 export * from './utils.js';
-export { PXEOracleInterface } from '../../../contract_function_simulator/pxe_oracle_interface.js';
 export type { PXECreationOptions } from '../../pxe_creation_options.js';
