+++ yarn-project/pxe/src/entrypoints/server/index.ts
@@ -3,6 +3,6 @@ export * from '../../config/index.js';
 export * from '../../error_enriching.js';
 export * from '../../storage/index.js';
 export * from './utils.js';
-export { PXEOracleInterface } from '../../contract_function_simulator/pxe_oracle_interface.js';
+export { NoteService } from '../../notes/note_service.js';
 export { ORACLE_VERSION } from '../../oracle_version.js';
 export { type PXECreationOptions } from '../pxe_creation_options.js';
