+++ yarn-project/pxe/src/notes/index.ts
@@ -1,0 +1,1 @@
+export { NoteService } from './note_service.js';
