+++ yarn-project/txe/src/txe_session.ts
@@ -8,7 +8,7 @@ import {
   AddressDataProvider,
   CapsuleDataProvider,
   NoteDataProvider,
-  PXEOracleInterface,
+  NoteService,
   PrivateEventDataProvider,
   RecipientTaggingDataProvider,
   SenderTaggingDataProvider,
@@ -121,14 +121,17 @@ export class TXESession implements TXESessionStateHandler {
       | IAvmExecutionOracle
       | ITxeExecutionOracle,
     private contractDataProvider: TXEContractDataProvider,
+    private noteDataProvider: NoteDataProvider,
     private keyStore: KeyStore,
     private addressDataProvider: AddressDataProvider,
-    private privateEventDataProvider: PrivateEventDataProvider,
     private accountDataProvider: TXEAccountDataProvider,
+    private senderTaggingDataProvider: SenderTaggingDataProvider,
+    private recipientTaggingDataProvider: RecipientTaggingDataProvider,
+    private capsuleDataProvider: CapsuleDataProvider,
+    private privateEventDataProvider: PrivateEventDataProvider,
     private chainId: Fr,
     private version: Fr,
     private nextBlockTimestamp: bigint,
-    private pxeOracleInterface: PXEOracleInterface,
   ) {}
 
   static async init(protocolContracts: ProtocolContract[]) {
@@ -156,27 +159,17 @@ export class TXESession implements TXESessionStateHandler {
     const version = new Fr(await stateMachine.node.getVersion());
     const chainId = new Fr(await stateMachine.node.getChainId());
 
-    const pxeOracleInterface = new PXEOracleInterface(
-      stateMachine.node,
-      keyStore,
-      contractDataProvider,
-      noteDataProvider,
-      capsuleDataProvider,
-      stateMachine.anchorBlockDataProvider,
-      senderTaggingDataProvider,
-      recipientTaggingDataProvider,
-      addressDataProvider,
-      privateEventDataProvider,
-    );
-
     const topLevelOracleHandler = new TXEOracleTopLevelContext(
       stateMachine,
       contractDataProvider,
+      noteDataProvider,
       keyStore,
       addressDataProvider,
-      privateEventDataProvider,
       accountDataProvider,
-      pxeOracleInterface,
+      senderTaggingDataProvider,
+      recipientTaggingDataProvider,
+      capsuleDataProvider,
+      privateEventDataProvider,
       nextBlockTimestamp,
       version,
       chainId,
@@ -189,14 +182,17 @@ export class TXESession implements TXESessionStateHandler {
       stateMachine,
       topLevelOracleHandler,
       contractDataProvider,
+      noteDataProvider,
       keyStore,
       addressDataProvider,
-      privateEventDataProvider,
       accountDataProvider,
+      senderTaggingDataProvider,
+      recipientTaggingDataProvider,
+      capsuleDataProvider,
+      privateEventDataProvider,
       version,
       chainId,
       nextBlockTimestamp,
-      pxeOracleInterface,
     );
   }
 
@@ -256,11 +252,14 @@ export class TXESession implements TXESessionStateHandler {
     this.oracleHandler = new TXEOracleTopLevelContext(
       this.stateMachine,
       this.contractDataProvider,
+      this.noteDataProvider,
       this.keyStore,
       this.addressDataProvider,
-      this.privateEventDataProvider,
       this.accountDataProvider,
-      this.pxeOracleInterface,
+      this.senderTaggingDataProvider,
+      this.recipientTaggingDataProvider,
+      this.capsuleDataProvider,
+      this.privateEventDataProvider,
       this.nextBlockTimestamp,
       this.version,
       this.chainId,
@@ -282,7 +281,11 @@ export class TXESession implements TXESessionStateHandler {
     // we perform this. We therefore search for known nullifiers now, as otherwise notes that were nullified would not
     // be removed from the database.
     // TODO(#12553): make the synchronizer sync here instead and remove this
-    await this.pxeOracleInterface.syncNoteNullifiers(contractAddress);
+    await new NoteService(
+      this.noteDataProvider,
+      this.stateMachine.node,
+      this.stateMachine.anchorBlockDataProvider,
+    ).syncNoteNullifiers(contractAddress);
 
     // Private execution has two associated block numbers: the anchor block (i.e. the historical block that is used to
     // build the proof), and the *next* block, i.e. the one we'll create once the execution ends, and which will contain
@@ -312,7 +315,16 @@ export class TXESession implements TXESessionStateHandler {
       new HashedValuesCache(),
       noteCache,
       taggingIndexCache,
-      this.pxeOracleInterface,
+      this.contractDataProvider,
+      this.noteDataProvider,
+      this.keyStore,
+      this.addressDataProvider,
+      this.stateMachine.node,
+      this.stateMachine.anchorBlockDataProvider,
+      this.senderTaggingDataProvider,
+      this.recipientTaggingDataProvider,
+      this.capsuleDataProvider,
+      this.privateEventDataProvider,
     );
 
     // We store the note and tagging index caches fed into the PrivateExecutionOracle (along with some other auxiliary
@@ -358,7 +370,11 @@ export class TXESession implements TXESessionStateHandler {
     // we perform this. We therefore search for known nullifiers now, as otherwise notes that were nullified would not
     // be removed from the database.
     // TODO(#12553): make the synchronizer sync here instead and remove this
-    await this.pxeOracleInterface.syncNoteNullifiers(contractAddress);
+    await new NoteService(
+      this.noteDataProvider,
+      this.stateMachine.node,
+      this.stateMachine.anchorBlockDataProvider,
+    ).syncNoteNullifiers(contractAddress);
 
     const anchorBlockHeader = await this.stateMachine.anchorBlockDataProvider.getBlockHeader();
 
@@ -367,7 +383,16 @@ export class TXESession implements TXESessionStateHandler {
       [],
       [],
       anchorBlockHeader,
-      this.pxeOracleInterface,
+      this.contractDataProvider,
+      this.noteDataProvider,
+      this.keyStore,
+      this.addressDataProvider,
+      this.stateMachine.node,
+      this.stateMachine.anchorBlockDataProvider,
+      this.senderTaggingDataProvider,
+      this.recipientTaggingDataProvider,
+      this.capsuleDataProvider,
+      this.privateEventDataProvider,
     );
 
     this.state = { name: 'UTILITY' };
